requirements.txt
Django==3.1.1
gevent==20.9.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
celery==5.0.5
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-filter==2.4.0
colorlog==5.0.1
requests==2.25.1
django-redis-cache==3.0.0
aliyun-log-python-sdk==0.6.53
